+++ src/config.js
@@ -1,0 +1,7 @@
+module.exports = Object.assign({
+  host: process.env.HOST || 'localhost',
+  port: process.env.PORT,
+  app: {
+    title: 'webpack-serve'
+  }
+})